Restructured: game classes moved to classes
